move transform function to utils and remove transformation from dataProvider
