+++ src/react-admin/dataProvider.js
@@ -34,7 +34,6 @@ export default {
 
   create: async (resource, { data }) => {
     const url = `${resource}`;
-    await transformData(data);
     const {
       json: { [resource.slice(0, -1)]: newItem },
     } = await httpClient(url, {
@@ -47,7 +46,6 @@ export default {
   update: async (resource, params) => {
     const { id, data } = params;
     const url = `${resource}/${id}`;
-    await transformData(data);
     const {
       json: { [resource.slice(0, -1)]: updatedItem },
     } = await httpClient(url, {
@@ -70,20 +68,3 @@ export default {
     return { data };
   },
 };
-
-const transformData = async (data) => {
-  if (data.image) {
-    const { image } = data;
-    const base64Image = await convertFileToBase64(image);
-    data.image = base64Image;
-  }
-};
-
-const convertFileToBase64 = (file) =>
-  new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onload = () => resolve(reader.result);
-    reader.onerror = reject;
-
-    reader.readAsDataURL(file.rawFile);
-  });
